requirements.txt
playwright
requests
# Web uygulaması
fastapi
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
jinja2
python-multipart
